+++ src/app/dashboard/alerts/alerts.component.html
@@ -1,17 +1,4 @@
-<div class="contacts-meta">
-
-    <!-- <mat-form-field appearance="outline">
-        <mat-select
-            (selectionChange)="filterBySeverity($event.value)"
-            [(value)]="filterSeverity"
-            panelClass="my-select-panel"
-        >
-            <mat-option *ngFor="let item of severity" [value]="item.value">
-            {{ item.viewValue }}
-            </mat-option>
-        </mat-select>
-    </mat-form-field> -->
-    
+<div class="contacts-meta">   
     <div class="halfer">
         <div *ngIf="alerts" class="pie-chart">
             <span class="chart-tip">click chart to <br>filter severity</span>
@@ -49,7 +36,7 @@
         [dataSource]="dataSource"
         multiTemplateDataRows
         matSort
-        (matSortChange)="sortChanged($event)"
+      
         >
 
         <!-- Checkbox Column -->
@@ -77,11 +64,6 @@
             </td>
         </ng-container>
 
-        <!-- <ng-container matColumnDef="errorMessage">
-            <th mat-header-cell *matHeaderCellDef mat-sort-header>MESSAGE</th>
-            <td mat-cell *matCellDef="let element" @tableCells>{{element.errorMessage}}</td>
-        </ng-container> -->
-
         <ng-container matColumnDef="errorCategory">
             <th mat-header-cell *matHeaderCellDef mat-sort-header>CATEGORY</th>
             <td mat-cell *matCellDef="let element" @tableCells>{{ element.errorCategory }}</td>
@@ -96,9 +78,9 @@
         <ng-container matColumnDef="expandedDetail">
             <td mat-cell *matCellDef="let element" [attr.colspan]="displayedColumns.length">
             <div class="example-element-detail"
+                [class.expanded]="element.expanded"
                 [@detailExpand]="element.expanded ? 'expanded' : 'collapsed'"
                 >
-                <!-- [@detailExpand]="element == expandedElement ? 'expanded' : 'collapsed'" -->
                 <div class="example-element-description theme-alt-font1">
                     {{ element.longMessage }} {{ element.expanded }}
                 </div>
@@ -147,7 +129,6 @@
         [pageSize]="pageSize"
         [length]="totalRows"
      
-        (page)="pageChanged($event)"
         >
     </mat-paginator>
 </div>
