+++ src/app/dashboard/contacts/contacts.component.html
@@ -84,9 +84,9 @@
         <ng-container matColumnDef="expandedDetail">
             <td mat-cell *matCellDef="let element" [attr.colspan]="displayedColumns.length">
             <div class="example-element-detail"
+                [class.expanded]="element.expanded"
                 [@detailExpand]="element.expanded ? 'expanded' : 'collapsed'"
                 >
-                <!-- [@detailExpand]="element == expandedElement ? 'expanded' : 'collapsed'" -->
 
                 <div class="example-element-description theme-alt-font1">
                     {{ element.contactDetail }}
@@ -122,7 +122,6 @@
             [class.example-expanded-row]="expandedElement === element"
             (click)="expander(element)"
             >
-            <!-- (click)="expandedElement = expandedElement === element ? null : element" -->
         </tr>
         <tr mat-row *matRowDef="let row; columns: ['expandedDetail']" class="example-detail-row"></tr>
         <tr mat-footer-row *matFooterRowDef="['loading']" [ngClass]="{'hide':dataSource!=null}"></tr>
